tidy and refactor
