--- src/main/kotlin/uk/gov/justice/digital/hmpps/subjectaccessrequesthtmlrenderer/controller/entity/SubjectDataHeldResponse.kt
@@ -1,8 +1,0 @@
-package uk.gov.justice.digital.hmpps.subjectaccessrequesthtmlrenderer.controller.entity
-
-data class SubjectDataHeldResponse(
-  val nomisId: String?,
-  val ndeliusId: String?,
-  val dataHeld: Boolean,
-  val serviceName: String?,
-)
