+++ src/main/kotlin/uk/gov/justice/digital/hmpps/subjectaccessrequesthtmlrenderer/service/DataHeldService.kt
@@ -1,11 +1,10 @@
 package uk.gov.justice.digital.hmpps.subjectaccessrequesthtmlrenderer.service
 
-import org.slf4j.LoggerFactory
 import org.springframework.http.HttpStatus
 import org.springframework.stereotype.Service
 import uk.gov.justice.digital.hmpps.subjectaccessrequesthtmlrenderer.client.DynamicServicesClient
+import uk.gov.justice.digital.hmpps.subjectaccessrequesthtmlrenderer.controller.BacklogController.SubjectDataHeldRequest
 import uk.gov.justice.digital.hmpps.subjectaccessrequesthtmlrenderer.controller.entity.RenderRequest
-import uk.gov.justice.digital.hmpps.subjectaccessrequesthtmlrenderer.controller.entity.SubjectDataHeldRequest
 import java.util.UUID
 
 @Service
@@ -13,23 +12,12 @@ class DataHeldService(
   private val dynamicServicesClient: DynamicServicesClient,
 ) {
 
-  private companion object {
-    private val log = LoggerFactory.getLogger(this::class.java)
-  }
-
   fun isSubjectDataHeld(request: SubjectDataHeldRequest): Boolean {
-    log.info(
-      "checking service={} for data held on nomisId:{} / ndeliusId:{}",
-      request.service!!.name,
-      request.nomisId,
-      request.ndeliusId,
-    )
-
     val response = dynamicServicesClient.getSubjectAccessRequestData(
       RenderRequest(
         id = UUID.randomUUID(),
-        serviceUrl = request.service.url,
-        serviceName = request.service.name,
+        serviceUrl = request.serviceUrl,
+        serviceName = request.serviceName,
         nomisId = request.nomisId,
         ndeliusId = request.ndeliusId,
         dateFrom = request.dateFrom,
@@ -37,14 +25,6 @@ class DataHeldService(
       ),
     )
 
-    return response?.statusCode == HttpStatus.OK.also {
-      log.info(
-        "data held={} for service={}, nomisId:{} / ndeliusId:{} ",
-        it,
-        request.service.name,
-        request.nomisId,
-        request.ndeliusId,
-      )
-    }
+    return response?.statusCode == HttpStatus.OK
   }
 }
